chore(gcp): Outputs options in the provider
